refactor: Split constants of String extension
* Rename String+Extension -> String+Constants
