+++ Expo1900/Expo1900/Extension/String+Constants.swift
@@ -1,0 +1,17 @@
+//
+//  String+Constants.swift
+//  Expo1900
+//
+//  Created by kjs on 2021/07/07.
+//
+
+import Foundation
+
+extension String {
+    static let zero = "0"
+    static let colonAndSpace = ": "
+    static let blank = ""
+    static let bracket = "("
+    static let newLine = "\n"
+    static let pageError = "페이지 오류"
+}
