--- Expo1900/Expo1900/Extension/String+Extension.swift
@@ -1,20 +1,0 @@
-//
-//  String+Extension.swift
-//  Expo1900
-//
-//  Created by kjs on 2021/07/07.
-//
-
-import Foundation
-
-extension String {
-    static let expositionFileName = "exposition_universelle_1900"
-    static let itemsFileName = "items"
-    static let bracket = "("
-    static let newLine = "\n"
-    static let zero = "0"
-    static let colon = ": "
-    static let blank = ""
-    static let people = " 명"
-    static let pageError = "페이지 오류"
-}
